+++ src/main/dll/SB/dll_0259_sbcloudrunner.c
@@ -605,7 +605,7 @@ void SB_CloudRunner_init(GameObject* obj)
     state->texture0 = textureLoadAsset(342);
     state->texture1 = textureLoadAsset(3085);
     *(void**)&state->resource = Resource_Acquire(121, 1);
-    ObjHits_SetTargetMask((int)obj, 1);
+    ObjHits_SetTargetMask(obj, 1);
     ObjGroup_AddObject((int)obj, SBCLOUDRUNNER_OBJGROUP);
 }
